requirements.txt
django>=3.1.1,<3.2.0
djangorestframework==3.11.1
psycopg2>=2.7.5,<2.8.0
flake8>=3.6.0,<3.7.0
